Refactor StocksPage to utilize useCalendarControls hook for improved date management
- Integrated useCalendarControls hook to manage calendar state and navigation, simplifying date handling in StocksPage.
- Removed redundant date management functions and replaced them with the hook's methods for better code clarity and maintainability.
- Updated the display of the current period to use the formatted output from the new hook, enhancing user experience.
